:people_holding_hands: Contributor Covenant: Update from v1.4 to v2.1
This commit updates the Contributor Covenant used for this repository's
Code of Conduct from v1.4 to v2.0. The most significant change between
the versions is the addition of a consequence ladder, or a way to
transparently share moderation guidelines and principles through the
Code of Conduct.

Although I hope to not have to use this, I believe the v2.0 changes are
helpful in providing structure and process in small communities like
ours where we have not fully explored what equitable consequences and
moderation best practices are.

CC: @jancborchardt @sleepypioneer @jonatoni @kristiprogri

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

:package: :arrow_up: Code of Conduct: Update to v2.1

Since this Pull Request was first opened, the Contributor Covenant was
updated to v2.1 with minor language changes. This commit incorporates
those changes and also makes the Code of Conduct respect the one
sentence per line convention, making diffs easier to review.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>