:alembic: :wrench: ci: Use tidy to catch HTML errors in CI (closes #120) (#122)
This commit removes the GitLab CI cruft and adds some Circle CI config
files for the site. Right now, the pipeline is really simple since our
site is really simple. It is just one raw HTML page.

This may change over time if we explore other tools like Jekyll or Hugo
for building and publishing a static website.

Closes #120. It is passing for an implementation, but the CI pipeline is
not as integrated as I would have liked. But it works and it enables us
to better catch problems before they make it to production.

CC: @jancborchardt @jonatoni

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>