Removing section from the website.
This resolves #115 by removing the "Your First PR" guide from the homepage. This was resolved in a PR by @everly-gif back on June 21, but there were merge conflicts (+ I don't have access to their branch, so this is simpler).

This should be good to go, tested it out locally and it looks great.